requirements.txt
requests == 2.32.3
dotenv == 0.9.9
pyyaml == 6.0.2